+++ Reldawin Unity/Assets/Scripts/XMLLoader/TETile.cs
@@ -5,7 +5,7 @@ using System.Xml.Serialization;
 public class TETile
 {
     public string name { get; set; }
-    public int id { get; set; }
+    public byte id { get; set; }
     public float minHeight { get; set; }
     public float maxHeight { get; set; }
     public int layerIndex { get; set; }
